% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{Create a density  plot of posterior distribution for a given parameter}
\usage{
density_plot(col, Mlist)
}
\arguments{
\item{col}{column that contains the parameter posterior draws}

\item{Mlist}{list that contains posterior draws from several gibbs sampler runs}
}
\description{
creates a density plot for a parameter
}
