% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_jeffrey.R
\name{gibbs_jeffreys}
\alias{gibbs_jeffreys}
\title{Runs a gibbs sampler chain with Jeffreys prior on N to estimate our model parameters}
\usage{
gibbs_jeffreys(seed, N = 10000, inclusionProbPriors)
}
\arguments{
\item{seed}{a seed}

\item{N}{number of iterations for the gibbs sampler}

\item{inclusionProbPriors}{priors for the inclusion probabilities of participating in each listing. should be present in a configureation file.}
}
\value{
a numeric matrix
}
\description{
Requires that parameters be declared globally, and note they are

hardcoded for this analysis
}
