% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_msm_binomial.R
\name{gibbs_msm_binomial}
\alias{gibbs_msm_binomial}
\title{Runs a gibbs sampler chain with a Binomial prior on N to estimate our model parameters}
\usage{
gibbs_msm_binomial(seed, N = 10000, atune = 0.25, btune = 1,
  inclusionProbPriors)
}
\arguments{
\item{seed}{a seed}

\item{N}{number of iterations for the gibbs sampler, default is 10,000}

\item{atune}{Tuning parameter for the metropolis-hastings update of hyperparameters on phis}

\item{btune}{Other tuning parameter for the metropolis-hastings update}

\item{inclusionProbPriors}{priors for the inclusion probabilities of participating in each listing. should be present in a configureation file.}

\item{x}{log of a.phi/(a.phi+ b.phi)}

\item{y}{log of (a.phi +b.phi)}

\item{phivec}{vector of current values of phi}
}
\value{
a numeric matrix
}
\description{
Requires that parameters be declared globally, and note they are

hardcoded for this analysis
}
