% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_msm_jeffrey.R
\name{gibbs_msm_jeffrey}
\alias{gibbs_msm_jeffrey}
\title{Runs a gibbs sampler chain with Jeffreys prior on N to estimate our model parameters}
\usage{
gibbs_msm_jeffrey(seed, N = 10000, inclusionProbPriors, gibbs.list, parnames)
}
\arguments{
\item{seed}{a seed}

\item{N}{number of iterations for the gibbs sampler, default is 10,000}

\item{inclusionProbPriors}{priors for the inclusion probabilities of participating in each listing. should be present in a configureation file.}

\item{gibbs.list}{list containing all the location listing data}

\item{parnames}{vector containing parameter names to be estimated}
}
\value{
a numeric matrix
}
\description{
Requires that parameters be declared globally, and note they are

hardcoded for this analysis
}
