% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_summary.R
\name{gibbs_summary}
\alias{gibbs_summary}
\title{Create summary output for Gibbs Sampler}
\usage{
gibbs_summary(x)
}
\arguments{
\item{x}{column of posterior draws for a parameter}
}
\description{
creates a trace plot for posterior draws
}
