% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln_petersen.R
\name{lincoln_petersen}
\alias{lincoln_petersen}
\title{Calculates the Lincoln-Petersen estimate, variance, and 95% CI}
\usage{
lincoln_petersen(n1, n2, m)
}
\arguments{
\item{n1}{the key pop count from the first survey (or listing)}

\item{n2}{the key pop count from the second survey (or listing)}

\item{m}{the count of individuals in both the first and second surveys or listings}
}
\value{
a list with element 1 the lincoln-petersen estimate, element 2 the variance, and element 3 the 95% confidence interval
}
\description{
Calculates the Lincoln-Petersen estimate, variance, and 95% CI
}
