% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_listing_count.R
\name{loc_listing_count}
\alias{loc_listing_count}
\title{Extract Number of participating individuals.}
\usage{
loc_listing_count(survLoc, listing, survData)
}
\arguments{
\item{survLoc}{character object with the corresponding location name in regionaldata.csv.}

\item{listing}{character object with the corresponding listing name in regionaldata.csv}

\item{survData}{csv file in the format of regionaldata.csv.}
}
\value{
a numeric object
}
\description{
Take regionaldata.csv file and get counts for

individuals who participated in specific listings.
}
