% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_city.R
\name{make_city}
\alias{make_city}
\title{Get the values for parameters for each location for use in simulation.}
\usage{
make_city(string, df)
}
\arguments{
\item{string}{The output of the Gibbs Chain contains parameter estimates labeled by location. This string corresponds to that location.}

\item{df}{a named vector or df that contains the parameter values}
}
\value{
a numeric vector
}
\description{
To simulate data we need values for parameters at each location
}
