% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_count.R
\name{overlap_count}
\alias{overlap_count}
\title{Get the counts of individuals who participated in multiple listings, when known.}
\usage{
overlap_count(survLoc, listing, overlapData)
}
\arguments{
\item{survLoc}{character object with the corresponding location name in overlapdata.csv.}

\item{listing}{character object with the corresponding listing name in overlapdata.csv}

\item{overlapData}{csv file in the format of overlapdata.csv.}
}
\value{
a numeric object
}
\description{
Take overlapdata.csv file and get counts for

individuals who participated in multiple listings.
}
