% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_msm_binomial.R
\name{sim_msm_binomial}
\alias{sim_msm_binomial}
\title{Simulates data for msm analysis and uses gibbs_msm_binomial
to evaluate how often the chain draws cover the true values}
\usage{
sim_msm_binomial(seed, N = 10000, inclusionProbPriors, atune = 0.25,
  btune = 1)
}
\arguments{
\item{seed}{a seed}

\item{N}{number of iterations for the gibbs sampler, default is 10,000}

\item{inclusionProbPriors}{priors for the inclusion probabilities of participating in each listing. should be present in a configureation file.}

\item{atune}{tuning parameter for metropolis-hastings step}

\item{btune}{tuning parameter for metropolis-hastings step}
}
\value{
a list that contains one iterations gibbs chain, a summary, and whether or not each parameter was recovered
}
\description{
Requires that parameters be declared globally, and note they are

hardcoded for this analysis
}
