% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_msm_jeffrey.R
\name{sim_msm_jeffrey}
\alias{sim_msm_jeffrey}
\title{Simulates data for msm analysis and uses gibbs_msm_jeffrey
to evaluate how often the chain draws cover the true values}
\usage{
sim_msm_jeffrey(seed, N = 10000, inclusionProbPriors)
}
\arguments{
\item{seed}{a seed}

\item{N}{number of iterations for the gibbs sampler, default is 10,000}

\item{inclusionProbPriors}{priors for the inclusion probabilities of participating in each listing. should be present in a configureation file.}
}
\value{
a list that contains one iterations gibbs chain, a summary, and whether or not each parameter was recovered
}
\description{
Requires that parameters be declared globally, and note they are

hardcoded for this analysis
}
