% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Create a trace plot of posterior draws for a given parameter}
\usage{
trace_plot(col, Mlist)
}
\arguments{
\item{col}{column that contains the parameter posterior draws}

\item{Mlist}{list that contains posterior draws from several gibbs sampler runs}
}
\description{
creates a trace plot for posterior draws
}
