% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ven_create.R
\name{ven_create}
\alias{ven_create}
\title{Creates a Venn Diagram with either  2 or 3 overlaps}
\usage{
ven_create(nListings, vecCounts, listNames, locName, groupColors)
}
\arguments{
\item{nListings}{number of listings}

\item{vecCounts}{a vector that contains the known number of participants for each partition of the Venn diagram}

\item{listNames}{a vector of the listing names for the legend}

\item{locName}{the location name for this Venn Diagram}

\item{groupColors}{color of the circle for each listing}
}
\value{
a ggplot object
}
\description{
Creates a listing Venn Diagram showing

missing and known data
}
